feat: ensure the refresh token belongs to a user

+++ src/presentation/middlewares/authentication/refresh-token/refresh-token-middleware.ts
@@ -1,7 +1,9 @@
+import { ComparerProtocol } from '@/application/protocols/cryptography'
 import { DecoderProtocol } from '@/application/protocols/cryptography/decoder-protocol'
 import { EncrypterProtocol } from '@/application/protocols/cryptography/encrypter-protocol'
 import { RefreshTokenDbRepositoryProtocol } from '@/application/protocols/repositories/refresh-token/refresh-token-db-repository-protocol'
-import { TokenMissingError } from '@/application/usecases/errors'
+import { UserDbRepositoryProtocol } from '@/application/protocols/repositories/user/user-db-repository-protocol'
+import { InvalidTokenError, TokenMissingError } from '@/application/usecases/errors'
 import {
   HttpHelperProtocol,
   HttpRequestToken,
@@ -10,10 +12,12 @@ import {
 import { MiddlewareProtocol } from '@/presentation/middlewares/protocols/middleware-protocol'
 export class RefreshTokenMiddleware implements MiddlewareProtocol {
   constructor (
+    private readonly comparer: ComparerProtocol,
     private readonly httpHelper: HttpHelperProtocol,
     private readonly jwtAccessToken: EncrypterProtocol,
     private readonly jwtRefreshToken: DecoderProtocol,
-    private readonly refreshTokenDbRepository: RefreshTokenDbRepositoryProtocol
+    private readonly refreshTokenDbRepository: RefreshTokenDbRepositoryProtocol,
+    private readonly userDbRepository: UserDbRepositoryProtocol
   ) {}
 
   async handle (httpRequest: HttpRequestToken): Promise<HttpResponse> {
@@ -30,6 +34,19 @@ export class RefreshTokenMiddleware implements MiddlewareProtocol {
     )
 
     if (!existentRefreshToken) return this.httpHelper.unauthorized(new TokenMissingError())
+
+    const refreshTokenOwner = await this.userDbRepository.findUserByRefreshToken(
+      existentRefreshToken.id
+    )
+
+    if (!refreshTokenOwner) return this.httpHelper.unauthorized(new TokenMissingError())
+
+    const isSameOwner = await this.comparer.compare(
+      refreshTokenOwner.personal.id,
+      existentRefreshToken.userId
+    )
+
+    if (!isSameOwner) return this.httpHelper.unauthorized(new InvalidTokenError())
 
     if (existentRefreshToken.expiresIn.getTime() < Date.now()) {
       await this.refreshTokenDbRepository.deleteRefreshTokenById(existentRefreshToken.userId)
